+++ tests/subsys/logging/log_msg2/src/main.c
@@ -308,7 +308,7 @@ void test_log_msg2_fp(void)
 	const void *source = (const void *)123;
 	int mode;
 	long long lli = 0x1122334455;
-	float f = 1.234;
+	float f = 1.234f;
 	double d = 11.3434;
 	char str[256];
 	int i = -100;
@@ -316,16 +316,16 @@ void test_log_msg2_fp(void)
 	test_init();
 
 	Z_LOG_MSG2_CREATE2(1, mode, 0, domain, source, level, NULL, 0,
-			TEST_MSG, i, lli, f, &i, d, source);
+			TEST_MSG, i, lli, (double)f, &i, d, source);
 	zassert_equal(mode, EXP_MODE(ZERO_COPY), NULL);
 
 	Z_LOG_MSG2_CREATE2(0, mode, 0, domain, source, level, NULL, 0,
-			TEST_MSG, i, lli, f, &i, d, source);
+			TEST_MSG, i, lli, (double)f, &i, d, source);
 	zassert_equal(mode, EXP_MODE(FROM_STACK), NULL);
 
 	z_log_msg2_runtime_create(domain, (void *)source, level, NULL, 0,
-				  TEST_MSG, i, lli, f, &i, d, source);
-	snprintfcb(str, sizeof(str), TEST_MSG, i, lli, f, &i, d, source);
+				  TEST_MSG, i, lli, (double)f, &i, d, source);
+	snprintfcb(str, sizeof(str), TEST_MSG, i, lli, (double)f, &i, d, source);
 
 	validate_base_message_set(source, domain, level,
 				   TEST_TIMESTAMP_INIT_VALUE,
